+++ src/_pytask/pluginmanager.py
@@ -49,6 +49,7 @@ def pytask_add_hooks(pm: PluginManager) -> None:
         "_pytask.dag_command",
         "_pytask.database",
         "_pytask.debugging",
+        "_pytask.explain",
         "_pytask.provisional",
         "_pytask.execute",
         "_pytask.live",