+++ CHANGELOG.md
@@ -8,6 +8,7 @@ releases are available on [PyPI](https://pypi.org/project/pytask) and
 ## 0.5.6 - 2025-xx-xx
 
 - {pull}`703` fixes {issue}`701` by allowing `--capture tee-sys` again.
+- {pull}`704` adds the `--explain` flag to show why tasks would be executed. Closes {issue}`466`.
 
 ## 0.5.5 - 2025-07-25